@@ -5,106 +5,108 @@
   "repository": "https://github.com/pankod/next-boilerplate",
   "author": "Pankod <info@pankod.com>",
   "scripts": {
-    "start:dev": "nodemon",
-    "build": "next build && tsc --project tsconfig.server.json",
-    "start": "cross-env NODE_ENV=production node dist/index.js",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "test": "jest --coverage",
     "lint": "eslint './{src,pages,server}/**/*' --ext .tsx,.ts",
     "pretty": "prettier --write --config .prettierrc.json {src,pages,server}/**/*.{ts,tsx}",
     "format": "npm run lint -- --fix",
-    "analyze": "cross-env BUNDLE_ANALYZE=both npm run build",
-    "export": "cross-env NODE_ENV=production next export",
+    "analyze": "next build && next analyze",
+    "export": "next export",
     "cli": "pankod-cli add",
     "storybook": "start-storybook -s ./public/static -p 6006",
     "build-storybook": "build-storybook -s ./public/static  -c .storybook -o .storybookout"
   },
   "dependencies": {
-    "@next/bundle-analyzer": "^9.4.4",
-    "autoprefixer": "^9.6.1",
-    "bootstrap": "^4.5.2",
-    "cross-env": "^5.2.1",
-    "dotenv": "^8.1.0",
-    "express": "^4.17.1",
-    "http-proxy-middleware": "^0.20.0",
+    "@next/bundle-analyzer": "^13.4.12",
+    "autoprefixer": "^10.4.14",
+    "bootstrap": "^5.3.1",
+    "cross-env": "^7.0.3",
+    "dotenv": "^16.3.1",
+    "eslint-config-next": "^13.4.12",
+    "express": "^4.18.2",
+    "http-proxy-middleware": "^2.0.6",
     "i18next-xhr-backend": "^3.2.2",
-    "isomorphic-unfetch": "^3.0.0",
-    "next": "^9.5.3",
-    "next-compose-plugins": "^2.2.0",
-    "next-i18next": "^4.5.0",
-    "next-redux-wrapper": "^3.0.0",
+    "isomorphic-unfetch": "^4.0.2",
+    "next": "^13.4.12",
+    "next-compose-plugins": "^2.2.1",
+    "next-i18next": "^14.0.0",
+    "next-redux-wrapper": "^8.1.0",
     "next-routes": "^1.4.2",
-    "query-string": "^6.8.3",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.1.1",
-    "redux": "^4.0.4",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.26.11",
-    "styled-components": "^4.3.2"
+    "query-string": "^8.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "redux": "^4.2.1",
+    "redux-thunk": "^2.4.2",
+    "sass": "^1.64.2",
+    "styled-components": "^6.0.7"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.0",
-    "@pankod/pankod-cli": "^1.0.0",
-    "@storybook/addon-actions": "^5.3.18",
-    "@storybook/addon-knobs": "^5.3.18",
-    "@storybook/addon-links": "^5.3.18",
-    "@storybook/addon-storysource": "^5.3.18",
-    "@storybook/addon-viewport": "^5.3.18",
-    "@storybook/addons": "^5.3.18",
+    "@babel/core": "^7.22.9",
+    "@pankod/pankod-cli": "^1.0.1",
+    "@storybook/addon-actions": "^7.2.1",
+    "@storybook/addon-knobs": "^7.0.2",
+    "@storybook/addon-links": "^7.2.1",
+    "@storybook/addon-storysource": "^7.2.1",
+    "@storybook/addon-viewport": "^7.2.1",
+    "@storybook/addons": "^7.2.1",
     "@storybook/preset-typescript": "^3.0.0",
-    "@storybook/react": "^5.3.18",
-    "@storybook/source-loader": "^5.3.18",
-    "@testing-library/jest-dom": "^5.5.0",
-    "@testing-library/react": "^10.0.4",
+    "@storybook/react": "^7.2.1",
+    "@storybook/source-loader": "^7.2.1",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
     "@types/dotenv": "^6.1.1",
-    "@types/express": "^4.17.1",
-    "@types/jest": "^24.0.18",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.3",
     "@types/next-redux-wrapper": "^2.0.2",
     "@types/nock": "^10.0.3",
-    "@types/node": "^12.7.5",
-    "@types/react": "^16.9.2",
-    "@types/react-redux": "^7.1.2",
-    "@types/redux-mock-store": "^1.0.1",
-    "@types/styled-components": "^4.1.19",
+    "@types/node": "^20.4.7",
+    "@types/react": "^18.2.18",
+    "@types/react-redux": "^7.1.25",
+    "@types/redux-mock-store": "^1.0.3",
+    "@types/styled-components": "^5.1.26",
     "@types/testing-library__react": "^10.0.1",
-    "@typescript-eslint/eslint-plugin": "^2.3.0",
-    "@typescript-eslint/parser": "^2.30.0",
+    "@typescript-eslint/eslint-plugin": "^6.2.1",
+    "@typescript-eslint/parser": "^6.2.1",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "10.x",
-    "babel-jest": "^25.5.1",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-module-resolver": "^3.2.0",
-    "babel-plugin-styled-components": "^1.10.6",
-    "babel-preset-react-app": "^9.1.2",
-    "css-loader": "^3.5.1",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-config-airbnb-typescript": "^4.0.1",
-    "eslint-config-prettier": "^6.3.0",
-    "eslint-config-react-app": "^5.0.2",
+    "babel-jest": "^29.6.2",
+    "babel-loader": "^9.1.3",
+    "babel-plugin-module-resolver": "^5.0.0",
+    "babel-plugin-styled-components": "^2.1.4",
+    "babel-preset-react-app": "^10.0.1",
+    "css-loader": "^6.8.1",
+    "eslint": "^8.46.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-config-react-app": "^7.0.1",
     "eslint-import-resolver-ts": "^0.4.2",
-    "eslint-import-resolver-typescript": "^1.1.1",
-    "eslint-plugin-flowtype": "3.x",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest-dom": "^2.1.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-flowtype": "8.x",
+    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-jest-dom": "^5.0.1",
     "eslint-plugin-jsx-a11y": "6.x",
-    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react": "7.x",
-    "eslint-plugin-react-hooks": "1.x",
-    "fork-ts-checker-webpack-plugin": "^4.1.2",
-    "jest": "^25.5.3",
-    "jest-styled-components": "^6.3.4",
-    "nock": "^11.3.5",
-    "nodemon": "^2.0.3",
-    "prettier": "^1.19.1",
+    "eslint-plugin-react-hooks": "4.x",
+    "fork-ts-checker-webpack-plugin": "^8.0.0",
+    "jest": "^29.6.2",
+    "jest-styled-components": "^7.1.1",
+    "nock": "^13.3.2",
+    "nodemon": "^3.0.1",
+    "prettier": "^3.0.1",
     "redux-devtools-extension": "^2.13.8",
-    "redux-mock-store": "^1.5.3",
-    "storybook-addon-styled-component-theme": "^1.3.0",
-    "style-loader": "^1.1.3",
-    "ts-jest": "^25.4.0",
-    "ts-loader": "^6.2.2",
-    "ts-node": "^8.4.1",
-    "typescript": "^3.9.3"
+    "redux-mock-store": "^1.5.4",
+    "storybook-addon-styled-component-theme": "^2.0.0",
+    "style-loader": "^3.3.3",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.4.4",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.6"
   },
   "pankod": {
     "projectType": "nextjs2"
