@@ -5,50 +5,47 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const React = require('react');
+import React from 'react';
 
-const CompLibrary = require('../../core/CompLibrary.js');
-
-const Container = CompLibrary.Container;
-const GridBlock = CompLibrary.GridBlock;
+import { Container, GridBlock } from '../../core/CompLibrary.js';
 
 function Help(props) {
-	const { config: siteConfig, language = '' } = props;
-	const { baseUrl, docsUrl } = siteConfig;
-	const docsPart = `${docsUrl ? `${docsUrl}/` : ''}`;
-	const langPart = `${language ? `${language}/` : ''}`;
-	const docUrl = doc => `${baseUrl}${docsPart}${langPart}${doc}`;
+  const { config: siteConfig, language = '' } = props;
+  const { baseUrl, docsUrl } = siteConfig;
+  const docsPart = `${docsUrl ? `${docsUrl}/` : ''}`;
+  const langPart = `${language ? `${language}/` : ''}`;
+  const docUrl = doc => `${baseUrl}${docsPart}${langPart}${doc}`;
 
-	const supportLinks = [
-		{
-			content: `Learn more using the [documentation on this site.](${docUrl(
-				'doc1.html',
-			)})`,
-			title: 'Browse Docs',
-		},
-		{
-			content: 'Ask questions about the documentation and project',
-			title: 'Join the community',
-		},
-		{
-			content: "Find out what's new with this project",
-			title: 'Stay up to date',
-		},
-	];
+  const supportLinks = [
+    {
+      content: `Learn more using the [documentation on this site.](${docUrl(
+        'doc1.html',
+      )})`,
+      title: 'Browse Docs',
+    },
+    {
+      content: 'Ask questions about the documentation and project',
+      title: 'Join the community',
+    },
+    {
+      content: "Find out what's new with this project",
+      title: 'Stay up to date',
+    },
+  ];
 
-	return (
-		<div className="docMainWrapper wrapper">
-			<Container className="mainContainer documentContainer postContainer">
-				<div className="post">
-					<header className="postHeader">
-						<h1>Need help?</h1>
-					</header>
-					<p>This project is maintained by a dedicated group of people.</p>
-					<GridBlock contents={supportLinks} layout="threeColumn" />
-				</div>
-			</Container>
-		</div>
-	);
+  return (
+    <div className="docMainWrapper wrapper">
+      <Container className="mainContainer documentContainer postContainer">
+        <div className="post">
+          <header className="postHeader">
+            <h1>Need help?</h1>
+          </header>
+          <p>This project is maintained by a dedicated group of people.</p>
+          <GridBlock contents={supportLinks} layout="threeColumn" />
+        </div>
+      </Container>
+    </div>
+  );
 }
 
-module.exports = Help;
+export default Help;
