@@ -1,13 +1,5 @@
-/**
- * Copyright (c) 2017-present, Facebook, Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-const React = require('react');
-
-const CompLibrary = require('../../core/CompLibrary.js');
+import React from "react";
+import CompLibrary from '../../core/CompLibrary.js';
 
 const Container = CompLibrary.Container;
 
@@ -45,4 +37,4 @@ class Users extends React.Component {
 	}
 }
 
-module.exports = Users;
+export default Users;
