@@ -73,7 +73,7 @@ describe("Home action tests", () => {
         ];
 
         // eslint-disable-next-line
-        await store.dispatch<any>(
+        (await store.dispatch)<any>(
             HomeActions.GetApod({ params: { hd: true } })
         );
 
